requirements.txt
Pillow==9.4.0
Pillow==10.1.0
Requests==2.31.0
streamlit==1.17.0
streamlit_lottie==0.0.3
